refactor: Controller 계층과 동일한 dto 통합
- 현재 규모가 작은 프로젝트라 Service 계층의 dto를 공통적으로 쓰도록 변경함
