--- src/main/java/roomescape/controller/dto/request/ReservationTimeCreateRequest.java
@@ -1,14 +1,0 @@
-package roomescape.controller.dto.request;
-
-import roomescape.service.dto.request.ReservationTimeData;
-
-import java.time.LocalTime;
-
-public record ReservationTimeCreateRequest(
-        LocalTime startAt
-) {
-
-    public ReservationTimeData toData() {
-        return new ReservationTimeData(startAt);
-    }
-}
