+++ src/main/java/roomescape/service/ReservationService.java
@@ -7,8 +7,8 @@ import roomescape.domain.ReservationTime;
 import roomescape.domain.dto.ReservationCreateCommand;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
-import roomescape.service.dto.request.ReservationCreateData;
-import roomescape.service.dto.response.ReservationResult;
+import roomescape.service.dto.request.ReservationCreateRequest;
+import roomescape.service.dto.response.ReservationResponse;
 
 import java.util.List;
 
@@ -19,14 +19,14 @@ public class ReservationService {
     private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
 
-    public List<ReservationResult> getReservations() {
+    public List<ReservationResponse> getReservations() {
         return reservationRepository.findAll()
                 .stream()
-                .map(ReservationResult::from)
+                .map(ReservationResponse::from)
                 .toList();
     }
 
-    public ReservationResult create(final ReservationCreateData data) {
+    public ReservationResponse create(final ReservationCreateRequest data) {
         final ReservationTime reservationTime = reservationTimeRepository.findById(data.timeId())
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 예약 시간입니다."));
 
@@ -40,7 +40,7 @@ public class ReservationService {
 
         final Reservation savedReservation = reservationRepository.save(reservation);
 
-        return ReservationResult.from(savedReservation);
+        return ReservationResponse.from(savedReservation);
     }
 
     public void delete(final Long reservationId) {
