+++ src/main/java/roomescape/service/dto/response/ReservationTimeResponse.java
@@ -4,13 +4,13 @@ import roomescape.domain.ReservationTime;
 
 import java.time.LocalTime;
 
-public record ReservationTimeResult(
+public record ReservationTimeResponse(
         Long id,
         LocalTime startAt
 ) {
 
-    public static ReservationTimeResult from(final ReservationTime time) {
-        return new ReservationTimeResult(
+    public static ReservationTimeResponse from(final ReservationTime time) {
+        return new ReservationTimeResponse(
                 time.getId(),
                 time.getStartAt()
         );
